+++ govtech_fasms/applicants/urls.py
@@ -5,7 +5,6 @@ from .views import ApplicantViewSet
 router = DefaultRouter()
 router.register('', ApplicantViewSet, basename='applicant')
 
-# URL patterns
 urlpatterns = [
     path('', include(router.urls)),
 ]